ai_api/requirements.txt
# Core AI and ML libraries for image tagging
# torch>=2.0.0
# torchvision>=0.15.0
# transformers>=4.30.0
torch>=2.2.0,<2.3.0
torchvision>=0.17.0,<0.22.0
transformers>=4.31.0,<4.50.0
ultralytics>=8.0.0
# yolov5>=7.0.0  # 移除，因为版本兼容性问题
# clip>=1.0  # 移除，因为包名错误
timm>=0.9.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
imageio>=2.30.0

# Data processing
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0
scikit-learn>=1.2.0
scikit-image>=0.20.0

# Web framework for API
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0
tqdm>=4.65.0

# Optional: for advanced features (uncomment if needed)
# matplotlib>=3.7.0
# seaborn>=0.12.0
# plotly>=5.15.0
